148: Use manifest placeholder to specify the module name
