+++ Library/src/main/AndroidManifest.xml
@@ -8,6 +8,10 @@
         <meta-data
             android:name="dev.testify.destination"
             android:value="@string/testifyDestination" />
+
+        <meta-data
+            android:name="dev.testify.module"
+            android:value="@string/testifyModule" />
     </application>
 
 </manifest>
