+++ Library/src/main/java/dev/testify/TestifyFeatures.kt
@@ -25,9 +25,7 @@
 package dev.testify
 
 import android.content.Context
-import android.content.pm.PackageManager
-import android.os.Bundle
-import androidx.annotation.VisibleForTesting
+import dev.testify.internal.helpers.getMetaDataBundle
 
 enum class TestifyFeatures(internal val tags: List<String>, private val defaultValue: Boolean = false) {
 
@@ -88,14 +86,4 @@ enum class TestifyFeatures(internal val tags: List<String>, private val defaultV
             }
         }
     }
-}
-
-internal fun getMetaDataBundle(context: Context): Bundle? {
-    val applicationInfo = if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.TIRAMISU) {
-        context.packageManager?.getApplicationInfo(context.packageName, PackageManager.ApplicationInfoFlags.of(0))
-    } else {
-        @Suppress("DEPRECATION")
-        context.packageManager?.getApplicationInfo(context.packageName, PackageManager.GET_META_DATA)
-    }
-    return applicationInfo?.metaData
 }
