+++ Library/src/main/res/values/strings.xml
@@ -1,4 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <string name="testifyDestination" translatable="false">default</string>
+    <string name="testifyModule" translatable="false" />
 </resources>
