+++ Samples/Legacy/src/main/res/layout/view_client_details.xml
@@ -14,7 +14,7 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_margin="16dp"
-        android:textColor="@android:color/holo_red_dark"
+        android:textColor="#ff212b36"
         tools:text="Client since January 1, 1970" />
 
     <de.hdodenhof.circleimageview.CircleImageView
@@ -29,13 +29,13 @@
         tools:src="@drawable/avatar1" />
 
     <com.google.android.material.card.MaterialCardView
-        app:cardElevation="8dp"
         android:id="@+id/info_card"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginLeft="@dimen/card_spacing"
         android:layout_marginRight="@dimen/card_spacing"
-        android:minHeight="100dp">
+        android:minHeight="100dp"
+        app:cardElevation="8dp">
 
         <LinearLayout
             android:layout_width="match_parent"
